IntelliJ: automated inspection-driven fixes 1/ (#1781)
* AbstractionPrefixSet: simplify ifs then go ahead and fixup equals

* IntelliJ: fix redundant and unused throws

* IntelliJ: simplify streams

* IntelliJ: functional reference can be folded

* IntelliJ: default annotation return value

* IntelliJ: unnecessary unboxing

* IntelliJ: unnecessary boxing

* IntelliJ: remove excess type annotations

* IntelliJ: try-finally -> try-with-resources

* IntelliJ: anonymouse type has shorter Java8 alternative

* IntelliJ: unnecessary enum modifier

* IntelliJ: unnecessary interface modifier

* IntelliJ: unnecessary semicolon

* IntelliJ: simplify if statements

* IntelliJ: unnecessary null check before instanceof

* IntelliJ: lambda can be replaced by method reference

+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/SubRange.java
@@ -50,7 +50,7 @@ public final class SubRange implements Serializable, Comparable<SubRange> {
         throw new BatfishException("Invalid subrange end: \"" + parts[1] + "\"", e);
       }
     } else if (o instanceof Integer) {
-      int i = ((Integer) o).intValue();
+      int i = (Integer) o;
       _start = i;
       _end = i;
     } else {
